#1046 P1: relocate frame/mod.rs inline tests to frame/tests.rs
frame/mod.rs went from 6,435 raw LOC (1,993 prod / 4,442 tests) to
1,995 LOC by moving the entire `#[cfg(test)] mod tests { ... }` block
into a sibling `tests.rs` file via `#[path = "tests.rs"]`.

Pure file-organization split — no test bodies modified, no production
code touched. The test module's `use super::*` continues to resolve
against frame/mod.rs because `#[path]` only redirects the file
location, not the module's parent.

LOC distribution after:
- frame/mod.rs:    1,995 (was 6,435)
- frame/tests.rs:  4,443 (new — same content, dedented one level)

P2/P3 will sub-split tests.rs by category (nat44/nat64/gre/tcp_seg/
icmp/rewrite) once the test interactions stabilize. Shipping P1
standalone keeps the mod.rs win independent of any test reorganization
risk.

cargo test --release: 865 passed, 0 failed.
